refactor(test): test_tracker_behaviours had tests from a lot of different trackers, now in different test files
